+++ src/ui/svelte/weg/state/getters.svelte.ts
@@ -1,0 +1,64 @@
+import { invoke, SeelenCommand, SeelenEvent, Settings, subscribe, Widget } from "@seelen-ui/lib";
+import type { UserAppWindowColors } from "@seelen-ui/lib/types";
+import { lazyRune } from "libs/ui/svelte/utils";
+
+export const currentMonitorId = Widget.getCurrent().decoded.monitorId!;
+
+export const monitors = lazyRune(() => invoke(SeelenCommand.SystemGetMonitors));
+subscribe(SeelenEvent.SystemMonitorsChanged, monitors.setByPayload);
+
+export const players = lazyRune(() => invoke(SeelenCommand.GetMediaSessions));
+subscribe(SeelenEvent.MediaSessions, players.setByPayload);
+
+export const notifications = lazyRune(() => invoke(SeelenCommand.GetNotifications));
+subscribe(SeelenEvent.Notifications, notifications.setByPayload);
+
+export const mousePos = lazyRune(async () => {
+  const [x, y] = await invoke(SeelenCommand.GetMousePosition);
+  return { x, y };
+});
+subscribe(SeelenEvent.GlobalMouseMove, ({ payload: [x, y] }) => {
+  mousePos.value = { x, y };
+});
+
+export const trashBinInfo = lazyRune(() => invoke(SeelenCommand.GetTrashBinInfo));
+subscribe(SeelenEvent.TrashBinChanged, trashBinInfo.setByPayload);
+
+export const settings = lazyRune(() => Settings.getAsync());
+Settings.onChange((s) => (settings.value = s));
+
+export const selfWinId = lazyRune(() => invoke(SeelenCommand.GetSelfWindowId));
+
+export const interactables = lazyRune(() => invoke(SeelenCommand.GetUserAppWindows));
+subscribe(SeelenEvent.UserAppWindowsChanged, interactables.setByPayload);
+
+export const previews = lazyRune(() => invoke(SeelenCommand.GetUserAppWindowsPreviews));
+subscribe(SeelenEvent.UserAppWindowsPreviewsChanged, previews.setByPayload);
+
+export const windowsColors = lazyRune<Record<number, UserAppWindowColors>>(
+  () => invoke(SeelenCommand.GetUserAppWindowsColors),
+);
+subscribe(SeelenEvent.UserAppWindowsColorsChanged, windowsColors.setByPayload);
+
+export const focused = lazyRune(() => invoke(SeelenCommand.GetFocusedApp));
+
+export const widgetStatuses = lazyRune(() => invoke(SeelenCommand.DebugGetWidgetsStatuses));
+subscribe(SeelenEvent.WidgetDebugInfoChanged, widgetStatuses.setByPayload);
+
+export const wegItems = lazyRune(() => invoke(SeelenCommand.StateGetWegItems));
+
+await Promise.all([
+  monitors.init(),
+  players.init(),
+  notifications.init(),
+  mousePos.init(),
+  trashBinInfo.init(),
+  settings.init(),
+  selfWinId.init(),
+  interactables.init(),
+  previews.init(),
+  windowsColors.init(),
+  focused.init(),
+  widgetStatuses.init(),
+  wegItems.init(),
+]);